rozert-pay/pyproject.toml
[tool.poetry]
name = "rozert-pay"
version = "0.1.0"
description = ""
authors = ["Stanislav Kaledin <staskaledin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"
django = "^5.0.4"
celery = "^5.4.0"
pydantic = "^2.7.1"
psycopg2-binary = "^2.9.9"
cryptography = "==41.0.6"
pytz = "^2024.1"
django-webpack-loader = "^3.1.0"
sentry-sdk = "^2.9.0"
drf-spectacular = "^0.27.2"
django-object-actions = "^4.0.0"
django-json-widget = "==0.2.*"
mypy = "^1.10.0"
requests = "^2.32.3"
ipython = "^8.29.0"
python-dotenv = "^1.0.1"
factory-boy = "^3.3.0"
whitenoise = "^6.8.2"
gunicorn = "^23.0.0"
prometheus-client = "^0.21.1"
werkzeug = "^3.1.3"
flask = "^3.1.0"
prometheus-summary = "^0.1.3"
qrcode = "^8.0"
pyopenssl = "^23.2.0"
types-pyopenssl = "^23.1.0.2"
pillow = "=10.3.0"
django-celery-beat = "^2.7.0"
unidecode = "^1.4.0"
django-extensions = "^4.1"
django-redis = "^6.0.0"
slack-sdk = "^3.36.0"
types-pytz = "^2025.2.0.20250516"
ijson = "^3.4.0"
typing-extensions = "^4.14.1"
django-auditlog = "^3.2.1"
pycountry = "^24.6.1"
xmltodict = "^0.14.2"
PyJWT = "^2.8.0"
django-admin-rangefilter = "^0.13.3"
xlsxwriter = "^3.2.9"
paymentsds-mpesa = {git="https://github.com/sallyruthstruik/mpesa-python-sdk", rev="51d5294574d8e3ce26d02263b9e3a3e618f8ac2d"}
django-waffle = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
django-stubs = "^5.0.4"
djangorestframework-stubs = "^3.15.0"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"
pre-commit = "^3.7.0"
black = "23.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
celery-stubs = "^0.1.3"
requests-mock = "^1.12.1"
freezegun = "^1.5.1"
markupsafe = "2.1.5"
pylint = "^3.3.1"
pygithub = "^2.4.0"
diff-cover = "^9.2.0"
types-pytz = "^2025.2.0.20250516"
pytest-xdist = "^3.8.0"
pytest-mock = "^3.15.1"
types-xmltodict = "^0.15.0.20250907"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
